feat: 수익률 계산 기능 구현, 구입 금액의 자료형 long으로 변경
1. Person
- calculateReturnRate 메소드 추가
2. PurchaseAmount
- calculateReturnRate 메소드 추가
3. ReturnRateCalculator
- calculateReturnRate 기능 구현
4. 구입 금액의 자료형 long으로 변경
- 관련 메소드 수정
5. Rank
- getPrize 메소드 추가

+++ src/main/java/lotto/service/LottoListGenerator.java
@@ -7,8 +7,8 @@ import java.util.ArrayList;
 
 public class LottoListGenerator {
 
-    public ArrayList<Lotto> generateLottoList(int purchaseAmount, int MONEY_UNIT, LottoGenerator lottoGenerator) {
-        int lottoCount = purchaseAmount / MONEY_UNIT;
+    public ArrayList<Lotto> generateLottoList(long purchaseAmount, int MONEY_UNIT, LottoGenerator lottoGenerator) {
+        long lottoCount = purchaseAmount / MONEY_UNIT;
         ArrayList<Lotto> lottos = new ArrayList<>();
         for (int i = 0; i < lottoCount; i++) {
             lottos.add(lottoGenerator.generateLotto());
